% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyou_data2019.R
\docType{data}
\name{kajin}
\alias{kajin}
\title{'Kazin' list of Manyoshu data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 787 rows and 1 columns.
}
\usage{
kajin
}
\description{
'Kazin' list of Manyoshu data
}
\seealso{
Other manyou_data: 
\code{\link{jikou}},
\code{\link{manyou}}
}
\concept{manyou_data}
\keyword{datasets}
