% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyou_data2019.R
\name{manyou_data}
\alias{manyou_data}
\title{Manyoshu data}
\source{
\url{https://c-able.ne.jp/~y_mura/manyou_data2019.zip}
}
\description{
Text data of the Manyoshu revised by Yoshimura Makoto and his contributors.
}
\seealso{
manyou kajin jikou
}
